Edit file set up dataset to download MangaSegmentation

+++ environments/set_up_dataset.sh
@@ -9,25 +9,27 @@ fi
 
 export HF_TOKEN=$(python environments/get_hf_token.py)
 
-hf download hal-utokyo/Manga109 --repo-type dataset --include Manga109_released_2023_12_07.zip --local-dir ./data --token $HF_TOKEN
-unzip ./data/Manga109_released_2023_12_07.zip -d ./data
-rm ./data/Manga109_released_2023_12_07.zip
-rm -rf ./data/.cache
-rm -rf ./data/__MACOSX
-
-
-if [[ "$(pwd)" == *"/characters-and-dialouges-association-in-comics" ]] || [[ "$(pwd)" == *"/workspace" ]]; then
-    echo "You are in the characters-and-dialouges-association-in-comics directory or /workspace"
-
-    if [ ! -d ./data/Human_Annotate_300 ]; then
-        gdown --fuzzy "https://drive.google.com/file/d/1nekCTGInk57Oe2cZvOjO7RRkUpS2mevw/view?usp=sharing" -O ./data/Human_Annotate_300.zip 
-        unzip -o ./data/Human_Annotate_300.zip -d ./data/
-        rm ./data/Human_Annotate_300.zip
-    else
-        echo "./data/Human_Annotate_300 already exists, skipping download."
-    fi
-
+if [ ! -d ./data/MangaSegmentation ]; then
+    hf download MS92/MangaSegmentation --repo-type=dataset --local-dir ./data/MangaSegmentation --token $HF_TOKEN
+    rm -rf ./data/MangaSegmentation/.cache
 else
-    echo "You are NOT in the characters-and-dialouges-association-in-comics directory or /workspace"
-    echo "Change the directory to characters-and-dialouges-association-in-comics in local or /workspace in docker"
+    echo "./data/MangaSegmentation already exists, skipping download."
+fi
+
+if [ ! -d ./data/Manga109_released_2023_12_07 ]; then
+    hf download hal-utokyo/Manga109 --repo-type dataset --include Manga109_released_2023_12_07.zip --local-dir ./data --token $HF_TOKEN
+    unzip ./data/Manga109_released_2023_12_07.zip -d ./data
+    rm ./data/Manga109_released_2023_12_07.zip
+    rm -rf ./data/.cache
+    rm -rf ./data/__MACOSX
+else
+    echo "./data/Manga109_released_2023_12_07 already exists, skipping download."
+fi
+
+if [ ! -d ./data/Human_Annotate_300 ]; then
+    gdown --fuzzy "https://drive.google.com/file/d/1nekCTGInk57Oe2cZvOjO7RRkUpS2mevw/view?usp=sharing" -O ./data/Human_Annotate_300.zip 
+    unzip -o ./data/Human_Annotate_300.zip -d ./data/
+    rm ./data/Human_Annotate_300.zip
+else
+    echo "./data/Human_Annotate_300 already exists, skipping download."
 fi
